refactor(table): add reference implementation for generic tables
