+++ client/app/lib/components/core/table/Cell.tsx
@@ -6,23 +6,29 @@ interface CellProps {
   cell: CellData;
   align?: Alignment;
   className?: string;
+  header?: boolean;
 }
 
 const Cell = (props: CellProps): JSX.Element => {
-  const { cell, align, className } = props;
+  const { cell, align, className, header } = props;
 
   if (isCustomCell(cell))
     return (
       <MuiTableCell
         align={align}
         className={`${className ?? ''} ${cell.className ?? ''}`}
+        variant={header ? 'head' : 'body'}
       >
         {cell.render}
       </MuiTableCell>
     );
 
   return (
-    <MuiTableCell align={align} className={className ?? ''}>
+    <MuiTableCell
+      align={align}
+      className={className ?? ''}
+      variant={header ? 'head' : 'body'}
+    >
       {cell}
     </MuiTableCell>
   );
