+++ client/app/lib/components/core/table/index.tsx
@@ -1,4 +1,4 @@
-import Table from './Table';
-
-export default Table;
+export { default as HorizontalTable } from './HorizontalTable';
+export { default as TableContainer } from './TableContainer';
 export type { ColumnTemplate } from './template';
+export { default as VerticalTable } from './VerticalTable';
